requirements.txt
cassandra-driver
pyasyncore